#7255 Finish Privacy API implementation

+++ classes/privacy/provider.php
@@ -214,23 +214,30 @@ class provider implements
     public static function get_contexts_for_userid(int $userid): contextlist {
         $contextlist = new contextlist();
 
-        $params = ['userid' => $userid];
+        $params = [
+            'modulename' => 'assign',
+            'contextlevel' => CONTEXT_MODULE,
+            'userid' => $userid
+        ];
 
-        $sql = "SELECT c.id FROM {context} c
-                INNER JOIN {course_modules} cm ON cm.id = c.instanceid
+        $sql = "SELECT c.id 
+                FROM {context} c 
+                INNER JOIN {course_modules} cm ON cm.id = c.instanceid 
+                INNER JOIN {modules} m ON cm.module = m.id AND m.name = :modulename
                 INNER JOIN {local_assignsubm_download} d ON d.cmid = cm.id
-                WHERE d.userid = :userid";
+                WHERE ( d.userid = :userid
+                    AND c.contextlevel = :contextlevel )";
 
-        //$sql = "SELECT cmid FROM {local_assignsubm_download} WHERE userid = :userid";
         $contextlist->add_from_sql($sql, $params);
 
-        // $sql = "SELECT c.id FROM {context} c
-        //         INNER JOIN {course_modules} cm ON cm.id = c.instanceid
-        //         INNER JOIN {local_assignsubm_feedback} d ON d.cmid = cm.id
-        //         WHERE d.userid = :userid";
-
-        //$sql = "SELECT cmid FROM {local_assignsubm_feedback} WHERE userid = :userid";
-        //$contextlist->add_from_sql($sql, $params);
+        $sql = "SELECT c.id 
+                 FROM {context} c 
+                 INNER JOIN {course_modules} cm ON cm.id = c.instanceid 
+                 INNER JOIN {modules} m ON cm.module = m.id AND m.name = :modulename
+                 INNER JOIN {local_assignsubm_feedback} d ON d.cmid = cm.id
+                 WHERE ( d.userid = :userid
+                     AND c.contextlevel = :contextlevel )";
+        $contextlist->add_from_sql($sql, $params);
 
         return $contextlist;
     }
@@ -246,18 +253,21 @@ class provider implements
         if (empty($contextlist->count())) {
             return;
         }
-        // $context = $contextlist->get_contexts()[0];
+
         $userid = $contextlist->get_user()->id;
         foreach ($contextlist->get_contexts() as $context) {
-            $downloads = $DB->get_records('local_assignsubm_download', ['userid' => $userid]);
-            foreach ($downloads as $download) {
-                writer::with_context($context)->export_data([], $download);
+            $exportdata = new \stdClass();
+            $feedbacks = $DB->get_records('local_assignsubm_feedback', ['userid' => $userid, 'cmid' => $context->instanceid]);
+            foreach ($feedbacks as $feedback) {
+                $exportdata->feedback = $feedback;
             }
 
-            //$feedbacks = $DB->get_records('local_assignsubm_feedback', ['cmid' => $context->instanceid, 'userid' => $userid]);
-            //foreach ($feedbacks as $feedback) {
-            //    writer::with_context($context)->export_data([], $feedback);
-            //}
+            $downloads = $DB->get_records('local_assignsubm_download', ['userid' => $userid, 'cmid' => $context->instanceid]);
+            foreach ($downloads as $download) {
+                $exportdata->download = $download;
+            }
+
+            writer::with_context($context)->export_data([], $exportdata);
         }
     }
 
@@ -328,8 +338,13 @@ class provider implements
         $cmid = $context->instanceid;
         $userids = $userlist->get_userids();
 
-        $params = ['cmid' => $cmid, 'userids' => $userids];
-        $sql = "cmid = :cmid AND userid IN (:userids)";
+        if (empty($userids)) {
+            return;
+        }
+
+        list($insql, $params) = $DB->get_in_or_equal($userids, SQL_PARAMS_NAMED);
+        $params['cmid'] = $cmid;
+        $sql = "cmid = :cmid AND userid $insql";
 
         $DB->delete_records_select('local_assignsubm_download', $sql, $params);
         $DB->delete_records_select('local_assignsubm_feedback', $sql, $params);
